requirements.txt
streamlit==1.45.1
pandas
sqlalchemy
plotly
matplotlib
python-dotenv
rapidfuzz
openpyxl
xlsxwriter
requests
Pillow
python-dateutil
psycopg2-binary
psycopg2
streamlit-aggrid
streamlit-extras
streamlit-cookies-controller
mysql-connector-python
pymysql
apscheduler